\name{aggFeatures}
\alias{aggFeatures}
\title{aggFeatures}
\description{
Series of functions to Average over gene symbols, obtained from package pathRender
}

\usage{
aggFeatures(fvector, splitvar, aggfun)
}

\arguments{
  \item{fvector}{this is a full vector that needs splitting}
  \item{splitvar}{this is the covariate which needs splitting}
  \item{aggfun}{an internal function in Biobase that is required dependency}

}

\author{
Dr. Shahab Asgharzadeh \cr
Janahan Gnanachandran \cr
Anthony Colombo \cr
Alexander Magsarili \cr

}

