% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPalette.R
\name{colorPalette}
\alias{colorPalette}
\title{Has a list of color palettes used}
\usage{
colorPalette(n)
}
\arguments{
\item{n}{n is the name of the color palette that one needs}
}
\description{
uses color palettes of interest
}
\author{
Shahab Asgharzadeh
}
