% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSurvivalFrame.R
\name{createSurvivalFrame}
\alias{createSurvivalFrame}
\title{createSurvivalFrame}
\usage{
createSurvivalFrame(f.survfit)
}
\arguments{
\item{f.survfit}{dataframe that consist of survival information: overall
survival status and survial days}
}
\description{
Setup a KM frame in ggplot2.
}
\examples{

#survivalFrame <- createSurvivalFrame(survfit)
#survivalFrame

}
\author{
Dr. Shahab Asgharzadeh \cr Janahan Gnanachandran \cr Anthony Colombo

Shahab Asgharzadeh
}
