% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2es.R
\name{df2es}
\alias{df2es}
\title{df2es}
\usage{
df2es(es)
}
\arguments{
\item{es}{expression set}
}
\description{
Converts a dataframe to an expressionset. Assumptions, rows have the
covariate inforamtion in non-numeric values, these rows are converted to
covariate part
}
\examples{

#data_2_es <- df2es(eset)
#data_2_es

}
\author{
Shahab Asgharzadeh
}
