% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_factor.R
\name{df_factor}
\alias{df_factor}
\title{df_factor}
\usage{
df_factor(df)
}
\arguments{
\item{df}{data frame}
}
\description{
Takes a dataframe and makes each column a factor, returns a dataframe
}
\author{
Shahab Asgharzadeh
}
