% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es2df.R
\name{es2df}
\alias{es2df}
\title{es2df}
\usage{
es2df(es)
}
\arguments{
\item{es}{expression set}
}
\description{
Converting Expressionset to a dataframe
}
\examples{

#results <- es2df(eset)
#results

}
\references{
"An Introduction to Bioconductor's ExpressionSet Class" Seth
Falcon, Martin Morgan, and Robert Gentleman \cr 6 October, 2006; revised 9
February, 2007 \cr
}
\author{
Shahab Asgharzadeh
}
\keyword{~kwd1}
\keyword{~kwd2}
