% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es2esCor.R
\name{es2esCor}
\alias{es2esCor}
\title{es2esCor}
\usage{
es2esCor(es1, es2, features_es1 = NULL, corrtype = "spearman",
  plot = TRUE)
}
\arguments{
\item{es1}{first expression set}

\item{es2}{second expression set}

\item{features_es1}{select features}

\item{corrtype}{Options for this argument: pearson, spearman, kendall.
Recommended spearman.}

\item{plot}{Logical command where if TRUE, includes a plot}
}
\description{
Calculates correlation between two es data (have to be the same size and
obviously same rownames).
}
\examples{

#results_cor_obj <- es2esCor(eset1, eset2, features_es1 = NULL, corrtype = "spearman", plot = TRUE)
#results_cor_obj


}
\references{
"An Introduction to Bioconductor's ExpressionSet Class" \cr Seth
Falcon, Martin Morgan, and Robert Gentleman \cr 6 October, 2006; revised 9
February, 2007 \cr
}
\author{
Shahab Asgharzadeh
}
