% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es2GSEAformat.R
\name{es2GSEAformat}
\alias{es2GSEAformat}
\title{es2GSEAformat}
\usage{
es2GSEAformat(es, covar, exprfilename, classfilename)
}
\arguments{
\item{es}{expression set}

\item{covar}{The covar label must be in the phenotype data of es.}

\item{exprfilename}{expression set file name that requires \code{.txt}}

\item{classfilename}{class label file name that requires \code{.cls}}
}
\description{
Converts Expression set data to .txt format that can be used in Gene Set
Enrichment Analysis (GSEA), a format that is used for differential
expression analysis, from BROAD Institute. Creates the Class label file
(.cls) using the covar variable provided .
}
\examples{

# es2GSEAformat(eset, covar="features_of_interest", "filename.txt", "filename.cls")
# Writes to current working directory 
# getwd()
# Files will be on that location 


}
\references{
"An Introduction to Bioconductor's ExpressionSet Class" \cr Seth
Falcon, Martin Morgan, and Robert Gentleman \cr 6 October, 2006; revised 9
February, 2007 \cr
}
\author{
Shahab Asgharzadeh
}
