% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esAddPheno_OnevsOthers.R
\name{esAddPheno_OnevsOthers}
\alias{esAddPheno_OnevsOthers}
\title{esAddPheno_OnevsOthers}
\usage{
esAddPheno_OnevsOthers(es, covar)
}
\arguments{
\item{es}{expression set}

\item{covar}{covariate}
}
\description{
Function to take a covariate and build several others that will generate
multiple covariates with each covariate versus combine of others. The
returned es will have x number of new columns (x being number of variables
in covar). The labels will be combine_x1 (where x1 is name of the variable
in covar)
}
\examples{

#results <- esAddPheno_OnevsOthers(eset, covariate_of_interest)
#results

}
\references{
"An Introduction to Bioconductor's ExpressionSet Class" \cr Seth
Falcon, Martin Morgan, and Robert Gentleman \cr 6 October, 2006; revised 9
February, 2007 \cr
}
\author{
Shahab Asgharzadeh
}
