% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esAnova.R
\name{esAnova}
\alias{esAnova}
\title{esAnova}
\usage{
esAnova(es, covar, base = levels(covar)[1])
}
\arguments{
\item{es}{expression set object}

\item{covar}{is a variable to include in the ANOVA model y ~ covar}

\item{base}{indicates which of the covar levels should be used as the base
level for ANOVA - default is the first one. When calling the function, you
don't need to include the base}
}
\description{
Uses esApply function on the expressionset class. Currently implemented for
one factor model with 3 groups. Returns a table with x2, pvalues, mean & SE
of each group, x2, fold change (fc) of each pairwise analyses. Also returns
the score (fc * x2) and absolute of the score, it adds the rownames of the
expression matrix of the expressionset to the returned table.
}
\examples{

##Assuming an Expression set is created from the references, the function call is straight foward. ##\\cr
#results <- esAnova(eset, covar = eset$group_of_interest)
#results


}
\references{
"An Introduction to Bioconductor's ExpressionSet Class" \cr Seth
Falcon, Martin Morgan, and Robert Gentleman \cr 6 October, 2006; revised 9
February, 2007 \cr
}
\author{
Shahab Asgharzadeh
}
