% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esFilter.R
\name{esFilter}
\alias{esFilter}
\title{esFilter}
\usage{
esFilter(es, cvlimit = 0.5, cv_i = "over", exprlimit, exp_i = "over",
  pctsample = 1)
}
\arguments{
\item{es}{expression set}

\item{cvlimit}{coefficient of variance limit}

\item{cv_i}{coefficient of variance either 'over' or 'under' the index}

\item{exprlimit}{expression limit numeric value}

\item{exp_i}{expression index either 'over' or 'under' the index}

\item{pctsample}{percentage of samples}
}
\description{
Filter function to filter expression set class. Based on cv of each
expression (greater or less than) certain cv limit and percent of samples
with (greater or less than) certain expression value limit.
}
\examples{

#filter <- esFilter(eset, cvlimit = 0.5, cv_i = "over", exprlimit = 2, exp_i = "over", pctsample = 1)
#filter

}
\references{
"An Introduction to Bioconductor's ExpressionSet Class" \cr Seth
Falcon, Martin Morgan, and Robert Gentleman \cr 6 October, 2006; revised 9
February, 2007 \cr
}
\author{
Shahab Asgharzadeh
}
