\name{esfunction_ver413-package}
\alias{esfunction_ver413-package}
\alias{esfunction_ver413}
\docType{package}
\title{
Expression Set Function Version 413
}
\description{
List of esfunctions for TARGET_PROJECT
}
\details{
\tabular{ll}{
Package: \tab esfunction_ver413\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-08-11\cr
License: \tab GPL-3\cr
}
}
\author{
Dr.Shahab Asgharzadeh
Maintainer: Dr.Shahab Asgharzadeh <sasgharzadeh@chla.usc.edu>
}
\keyword{ package }

