% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esGetSymbol.R
\name{esGetSymbol}
\alias{esGetSymbol}
\title{esGetSymbol}
\usage{
esGetSymbol(annotatedlist)
}
\arguments{
\item{annotatedlist}{argument from an annotated list that has gene symbols
merged with transcript ids with underscore. For example, some esets will
have duplicate genes for various transcript ids, so by calling the function,
esAnnotate will output every gene including duplicates. In this case, the
row names of the expression set will be as follows "gene_probid".}
}
\description{
Get the symbols from a gene list that has been annotated (i.e.
GENE__ProbeId)
}
\examples{

#longnmaeset <- esAnnotate(eset, anno)
#names_for_cutting <- featureNames(longnmaeset)
#head(names_for_cutting) 
# OUTPUT:
# "gene1_transcriptid1" "gene2_transcriptid2" "gene3_transcriptid3" "gene4_transcriptid4" 
# "gene5_transcriptid5" "gene6_transcriptid6"

#genesymbols <-esGetSymbol(names_for_cutting)
#head(genesymbols)
# OUTPUT:
# "gene1" "gene2" "gene3" "gene4" "gene5" "gene6"




}
\references{
"An Introduction to Bioconductor's ExpressionSet Class" \cr Seth
Falcon, Martin Morgan, and Robert Gentleman \cr 6 October, 2006; revised 9
February, 2007 \cr
}
\author{
Shahab Asgharzadeh
}
