% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esKmeans.R
\name{esKmeans}
\alias{esKmeans}
\title{esKmeans}
\usage{
esKmeans(es, kmeans = 2)
}
\arguments{
\item{es}{expression set}

\item{kmeans}{number of clusters assumed and returns an expressionset object}
}
\description{
esKmeans function finds the kmeans clusters of a expressionset and adds the
cluster info to the covariate file
}
\examples{

#results<- esKmeans(eset, kmeans = 2)
#results 

}
\references{
"An Introduction to Bioconductor's ExpressionSet Class" \cr Seth
Falcon, Martin Morgan, and Robert Gentleman \cr 6 October, 2006; revised 9
February, 2007 \cr
}
\author{
Shahab Asgharzadeh
}
