% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esLoadGermanNbl_old.R
\name{esLoadGermanNbl_old}
\alias{esLoadGermanNbl_old}
\title{esLoadGermanNbl_old}
\usage{
esLoadGermanNbl_old(type = "tsr", load_target_subset = TRUE,
  excludeYprobes = TRUE)
}
\arguments{
\item{type}{only type tsr (transcript)}

\item{load_target_subset}{takes a different annotation r-object dataset}

\item{excludeYprobes}{removes probeset that belongs to chromosome y}
}
\description{
Automatically load German Neuroblastoma data
}
\examples{

#esLoadGermanNbl_old(type = "tsr", load_target_subset = TRUE, excludeYprobes = TRUE)


}
\references{
"An Introduction to Bioconductor's ExpressionSet Class" \cr Seth
Falcon, Martin Morgan, and Robert Gentleman \cr 6 October, 2006; revised 9
February, 2007 \cr
}
\author{
Shahab Asgharzadeh
}
