% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esLoadMedullo.R
\name{esLoadMedullo}
\alias{esLoadMedullo}
\title{esLoadMedullo}
\usage{
esLoadMedullo(type = "tsr")
}
\arguments{
\item{type}{loads transcripts without cell lines}
}
\description{
Automatically load medulloblastoma data, removes chr Y features
}
\examples{

#medullo <- esLoadMedullo(type="tsr")

}
\references{
"An Introduction to Bioconductor's ExpressionSet Class" \cr Seth
Falcon, Martin Morgan, and Robert Gentleman \cr 6 October, 2006; revised 9
February, 2007 \cr
}
\author{
Shahab Asgharzadeh
}
