% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esLoadU133NBL.R
\name{esLoadU133NBL}
\alias{esLoadU133NBL}
\title{esLoadU133NBL}
\usage{
esLoadU133NBL(type = "tsr", excludeYprobes = TRUE)
}
\arguments{
\item{type}{this is the type of load, tsr is default}

\item{excludeYprobes}{logical command for excluding certain probes in Y}
}
\description{
loaded data from U133 NBL
}
\author{
Shahab Asgharzadeh
}
