% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esNMF.R
\name{esNMF}
\alias{esNMF}
\title{esNMF}
\usage{
esNMF(es, k = 2, run = 100, nmfmethod = "brunet")
}
\arguments{
\item{es}{expression set with all positive values}

\item{k}{is the rank, which is a specification of a single numeric value.
The the maximum linear independent columns of matrix A. You can determine
how many factorization you want.}

\item{run}{number of runs to perform}

\item{nmfmethod}{Default method "brunet"}
}
\description{
NMF for expressionset, returns a covariate vector. Non-Negative Matrix
Factorization (NMF) factorize a matrix V into two non-negative matrices WH.
}
\note{
Load package NMF.
}
\examples{

#NMF <- esNMF(Annoset, k=2, run = 100, nmfmethod = "brunet")
#NMF 

}
\references{
"An Introduction to Bioconductor's ExpressionSet Class" \cr Seth
Falcon, Martin Morgan, and Robert Gentleman \cr 6 October, 2006; revised 9
February, 2007 \cr

\url{cran.r-project.org/web/packages/NMF/NMF.pdf}
}
\author{
Shahab Asgharzadeh
}
\keyword{~kwd1}
\keyword{~kwd2}
