% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esPathwayplot.R
\name{esPathwayplot}
\alias{esPathwayplot}
\title{esPathwayplot}
\usage{
esPathwayplot(esPathwaySig_results, covarcompare, fdrlevel = 0.05,
  labeltopsigpaths = c(1:5))
}
\arguments{
\item{esPathwaySig_results}{pathway analysis results from
\code{esPathwaySig}}

\item{covarcompare}{covariate comparison of interest}

\item{fdrlevel}{level of false discovery rate to be considered}

\item{labeltopsigpaths}{Number of top pathways could be labeled}
}
\description{
This will plot the pathway results based on chi-squared value and the
eucledian distance. Number of top pathways could be labeled with
\code{labeltopsigpaths} option.
}
\examples{

#esPathwayplot(esPathwaySig_results, covarcompare, fdrlevel = 0.05, labeltopsigpaths = c(1:5))

}
\references{
"An Introduction to Bioconductor's ExpressionSet Class" \cr Seth
Falcon, Martin Morgan, and Robert Gentleman \cr 6 October, 2006; revised 9
February, 2007 \cr
}
\author{
Shahab Asgharzadeh
}
\keyword{~kwd1}
\keyword{~kwd2}
