% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esPathwayRun.R
\name{esPathwayRun}
\alias{esPathwayRun}
\title{esPathwayRun}
\usage{
esPathwayRun(es, covar, nsim = 10, anovaresults, pathwaytype = "c2.cp.kegg")
}
\arguments{
\item{es}{expression set}

\item{covar}{covariate}

\item{nsim}{number of simulate. Default to 10. Typically run for 10,000.}

\item{anovaresults}{results from anova}

\item{pathwaytype}{pathway type, gene set of interest from BROAD Institute}
}
\description{
For a given pathway type this will load the curated msidb and call the
esPathwaySig function. This could be applied to any pathway files.
}
\examples{

#esPathwayRun(eset, covar="covariate_of_interest", nsim = 10, anovaresults, pathwaytype = "c2.cp.kegg")

}
\references{
"An Introduction to Bioconductor's ExpressionSet Class" \cr Seth
Falcon, Martin Morgan, and Robert Gentleman \cr 6 October, 2006; revised 9
February, 2007 \cr
}
\author{
Shahab Asgharzadeh
}
\keyword{~kwd1}
\keyword{~kwd2}
