% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esPathwayTop.R
\name{esPathwayTop}
\alias{esPathwayTop}
\title{esPathwayTop}
\usage{
esPathwayTop(esPathwaySig_results, covarcompare,
  headerprefix = "ecu_pw_chi_sd_", fdrlevel = 0.05)
}
\arguments{
\item{esPathwaySig_results}{pathway analysis results from
\code{esPathwaySig}}

\item{covarcompare}{covariate comparison of interest}

\item{headerprefix}{columns of interest to be chosen from pathway results}

\item{fdrlevel}{level of false discovery rate to be considered}
}
\description{
Finds Top scoring Pathways (using Score Calculated by multiplying pathway
eucledian distance and SD of Chi2.
}
\examples{

#esPathwayTop(esPathwaySig_results, covarcompare, headerprefix = "ecu_pw_chi_sd_", fdrlevel = 0.05)

}
\references{
"An Introduction to Bioconductor's ExpressionSet Class" \cr Seth
Falcon, Martin Morgan, and Robert Gentleman \cr 6 October, 2006; revised 9
February, 2007 \cr
}
\author{
Shahab Asgharzadeh
}
\keyword{~kwd1}
\keyword{~kwd2}
