% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esPcaplot.R
\name{esPcaplot}
\alias{esPcaplot}
\title{esPcaplot}
\usage{
esPcaplot(es, features = "", covar = "", labelvar = "", pcs = c(1, 2),
  shape = FALSE, kmcolors = colorPalette("kmPalette"), scale = TRUE,
  center = TRUE)
}
\arguments{
\item{es}{expression set}

\item{features}{features = set of featurenames of the expression set, if
left blank all are used}

\item{covar}{used to color the points in the PCA}

\item{labelvar}{used to label the points in the PCA}

\item{pcs}{used to state which principle components to plot}

\item{shape}{TRUE/FALSE. TRUE if you want to add shapes.}

\item{kmcolors}{Loads the color palette.}

\item{scale}{Logical command that scales the data. TRUE for scale.}

\item{center}{Logical command that centers the data. TRUE for center.}
}
\description{
esPCA plot of selected features of an expressionset class
}
\note{
For argument kmcolors, one can create their own color palette called
kmPalette <- c("red", "blue", "darkgreen", "black", "orange, "yellow4",
"brown", "purple", "grey", "turquoise"). The amount of colors must match the
number of levels.
}
\examples{

# kmcolor, look at the note
#esPcaplot(eset, covar="group_of_interest", shape = FALSE, kmcolors = colorPalette("kmPalette"))

}
\references{
"An Introduction to Bioconductor's ExpressionSet Class" \cr Seth
Falcon, Martin Morgan, and Robert Gentleman \cr 6 October, 2006; revised 9
February, 2007 \cr
}
\author{
Shahab Asgharzadeh
}
\keyword{~kwd1}
\keyword{~kwd2}
