% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esPctiles.R
\name{esPctiles}
\alias{esPctiles}
\title{esPctiles}
\usage{
esPctiles(eset)
}
\arguments{
\item{eset}{expression set}
}
\description{
Returns an expressionset class where the values have been converted to
percentiles i.e. Calculate empirical continous desnity function and create a
expression set based on that.
}
\examples{

#ecdf_on_assay_data <- esPctiles(eset)
#ecdf_on_assay_data

}
\references{
"An Introduction to Bioconductor's ExpressionSet Class" \cr Seth
Falcon, Martin Morgan, and Robert Gentleman \cr 6 October, 2006; revised 9
February, 2007 \cr
}
\author{
Shahab Asgharzadeh
}
\keyword{~kwd1}
\keyword{~kwd2}
