% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esReduce.R
\name{esReduce}
\alias{esReduce}
\title{esReduce}
\usage{
esReduce(es, annot)
}
\arguments{
\item{es}{expression set}

\item{annot}{is an annotation object (first column is the probe_id)}
}
\description{
Create a common gene expressionset
}
\examples{

#results <- esReduce(eset, anno_obj)
#results


}
\references{
"An Introduction to Bioconductor's ExpressionSet Class" \cr Seth
Falcon, Martin Morgan, and Robert Gentleman \cr 6 October, 2006; revised 9
February, 2007 \cr
}
\author{
hahab Asgharzadeh
}
\keyword{~kwd1}
\keyword{~kwd2}
