% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esReduce2.R
\name{esReduce2}
\alias{esReduce2}
\title{esReduce2}
\usage{
esReduce2(es, annovec, ann2featMap, pdvname = "symbol", collapseFun = NULL)
}
\arguments{
\item{es}{expression set}

\item{annovec}{annotation vector, a collection of annotations}

\item{ann2featMap}{is either an AtomicAnnDbBimap with keys in a and range in
f, or a list with element names in a and element values in f}

\item{pdvname}{gene symbol}

\item{collapseFun}{collapse function, a function that applys to}
}
\description{
ExpressionSet with featureNames
}
\examples{

#esReduce2(eset, annovec, ann2featMap, pdvname = "symbol", collapseFun = NULL)

}
\author{
Shahab Asgharzadeh
}
\keyword{~kwd1}
\keyword{~kwd2}
