% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featNamesFromAnno.R
\name{featNamesFromAnno}
\alias{featNamesFromAnno}
\title{featNamesFromAnno}
\usage{
featNamesFromAnno(annovec, ann2featMap)
}
\arguments{
\item{annovec}{annotation vector}

\item{ann2featMap}{annotation2map}
}
\description{
Feature names of an expression set from annotation
}
\author{
Shahab Asgharzadeh
}
\keyword{~kwd1}
\keyword{~kwd2}
