% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixpath.R
\name{fixpath}
\alias{fixpath}
\title{fixpath}
\usage{
fixpath(x)
}
\arguments{
\item{x}{this is the input}
}
\description{
this creates a path to the data
}
\author{
Shahab Asgharzadeh
}
\keyword{~kwd1}
\keyword{~kwd2}
