% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmfunction.R
\name{glmfunction}
\alias{glmfunction}
\title{glmfunction}
\usage{
glmfunction(y, var1, contrastmodel = NULL)
}
\arguments{
\item{y}{a vector of data}

\item{var1}{the covariate}

\item{contrastmodel}{NULL}
}
\description{
Perform differential expression analysis for all possible combination of
from a vector of data and computes the mean, fold change, p-value,
chi-squared values, and generates a score for ranking.
}
\examples{

#x <- c(2,3,4,7,9,4,10,6,4,2)
#y <- as.factor(c("a","b","c","a","b","c","b","a","c","c"))
#results <- glmfunction(x,y) 
#results 


}
\author{
Shahab Asgharzadeh
}
\keyword{~kwd1}
\keyword{~kwd2}
