% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_ni.R
\name{ni}
\alias{ni}
\title{ni}
\usage{
ni(x, y)
}
\arguments{
\item{x}{this is an input, where the usage takes whatever is not in x and is
in y}

\item{y}{the second compared input.}
}
\description{
Setting color pallette for ggplot and to setup colors for heatmap
}
\author{
Shahab Asgharzadeh
}
\keyword{~kwd1}
\keyword{~kwd2}
