% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryAE2.R
\name{queryAE2}
\alias{queryAE2}
\title{queryAE2}
\usage{
queryAE2(keywords = NULL, species = NULL)
}
\arguments{
\item{keywords}{NULL}

\item{species}{NULL}
}
\description{
Gets you the names of the arrays
}
\examples{

#queryAE2(keywords = NULL, species = NULL)

}
\author{
Shahab Asgharzadeh
}
\keyword{~kwd1}
\keyword{~kwd2}
