% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revAnnoEnv.R
\name{revAnnoEnv}
\alias{revAnnoEnv}
\title{revAnnoEnv}
\usage{
revAnnoEnv(env)
}
\arguments{
\item{env}{envelope}
}
\description{
this gives a revises the input.  It takes an input, creates a list of
objects, prints the names of the input,
}
\author{
Shahab Asgharzadeh
}
\keyword{~kwd1}
\keyword{~kwd2}
