% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tmp_path_data.R
\name{set_tmp_path_data}
\alias{set_tmp_path_data}
\title{Set_Tmp_Path_Data}
\usage{
set_tmp_path_data()
}
\description{
This is a function that creates a temporary path directory for retrieving
data, behind the scene task function
}
\author{
Shahab Asgharzadeh
}
\keyword{~kwd1}
\keyword{~kwd2}
