% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.clipboard.R
\name{write.clipboard}
\alias{write.clipboard}
\title{wirte.clipboard}
\usage{
write.clipboard(x, col.names = FALSE)
}
\arguments{
\item{x}{this writes an input ready to be copied}

\item{col.names}{this is a logical command}
}
\description{
this is for copying inputs
}
\keyword{~kwd1}
\keyword{~kwd2}
