\name{ni}
\alias{ni}
\title{ni}
\description{
Setting color pallette for ggplot and to setup colors for heatmap
}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
this is an input, where the usage takes whatever is not in x and is in y
}
  \item{y}{
the second compared input.  
}
}

\author{
Dr. Shahab Asgharzadeh \cr
Janahan Gnanachandran \cr
Anthony Colombo \cr
Alexander Magsarili \cr

}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line