% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggFeatures.R
\name{aggFeatures}
\alias{aggFeatures}
\title{aggFeatures}
\usage{
aggFeatures(fvector, splitvar, aggfun)
}
\arguments{
\item{fvector}{this is a full vector that needs splitting}

\item{splitvar}{this is the covariate which needs splitting}

\item{aggfun}{an internal function in Biobase that is required dependency}
}
\description{
Series of functions to Average over gene symbols, obtained from package
pathRender
}
\author{
Shahab Asgharzadeh
}
