\name{df2es}
\alias{df2es}
\title{df2es}
\description{
Converts a dataframe to an expressionset. Assumptions, rows have the covariate inforamtion in non-numeric values, these rows are converted to covariate part
}
\usage{
df2es(es)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{es}{expression set

}
}

\examples{
#data_2_es <- df2es(eset)
#data_2_es
}
