\name{es2df}
\alias{es2df}
\title{es2df}
\description{
Converting Expressionset to a dataframe
}
\usage{
es2df(es)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{es}{expression set

}
}

\references{
"An Introduction to Bioconductor's ExpressionSet Class"
Seth Falcon, Martin Morgan, and Robert Gentleman \cr
6 October, 2006; revised 9 February, 2007 \cr

}
\author{
Dr. Shahab Asgharzadeh \cr
Janahan Gnanachandran \cr
Anthony Colombo \cr
Alexander Magsarili \cr

}

\examples{
#results <- es2df(eset)
#results
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
