% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esDropPheno.R
\name{esDropPheno}
\alias{esDropPheno}
\title{esDropPheno}
\usage{
esDropPheno(es, column_number)
}
\arguments{
\item{es}{expression set}

\item{column_number}{specify a column number and removes it from the
dataframe}
}
\description{
Function to drop a column of the phenoData of expressionset
}
\examples{

# 5 is column number you want to delete
#neweset <- esDropPheno(eset, 5)
#neweset


}
\references{
"An Introduction to Bioconductor's ExpressionSet Class" \cr Seth
Falcon, Martin Morgan, and Robert Gentleman \cr 6 October, 2006; revised 9
February, 2007 \cr
}
\author{
Shahab Asgharzadeh
}
