\name{esLoadGermanNbl_old}
\alias{esLoadGermanNbl_old}
\title{esLoadGermanNbl_old
}
\description{
Automatically load German Neuroblastoma data
}
\usage{
esLoadGermanNbl_old(type = "tsr", load_target_subset = TRUE, excludeYprobes = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{type}{only type tsr (transcript)

}
  \item{load_target_subset}{takes a different annotation r-object dataset

}
  \item{excludeYprobes}{removes probeset that belongs to chromosome y 

}
}

\references{
"An Introduction to Bioconductor's ExpressionSet Class" \cr
Seth Falcon, Martin Morgan, and Robert Gentleman \cr
6 October, 2006; revised 9 February, 2007 \cr


}
\author{
Dr. Shahab Asgharzadeh \cr
Janahan Gnanachandran \cr
Anthony Colombo \cr
Alexander Magsarili \cr
}

\examples{
#esLoadGermanNbl_old(type = "tsr", load_target_subset = TRUE, excludeYprobes = TRUE)

}
