\name{esLoadMedullo}
\alias{esLoadMedullo}
\title{esLoadMedullo}
\description{
Automatically load medulloblastoma data, removes chr Y features
}
\usage{
esLoadMedullo(type = "tsr")
}

\arguments{
  \item{type}{loads transcripts without cell lines 

}
}

\references{
"An Introduction to Bioconductor's ExpressionSet Class" \cr
Seth Falcon, Martin Morgan, and Robert Gentleman \cr
6 October, 2006; revised 9 February, 2007 \cr


}
\author{
Dr. Shahab Asgharzadeh \cr
Janahan Gnanachandran \cr
Anthony Colombo \cr
Alexander Magsarili \cr
}

\examples{
#medullo <- esLoadMedullo(type="tsr")
}
