% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esLoadTarget.R
\name{esLoadTarget}
\alias{esLoadTarget}
\title{esLoadTarget}
\usage{
esLoadTarget(type = NULL, load_target_subset = TRUE,
  excludeYprobes = TRUE)
}
\arguments{
\item{type}{4 types: tsr (transcript), psr (probe script), avgaffy (average
transcript base on affy notation), avgref (average transcript based on ref
gene)}

\item{load_target_subset}{Logical argument. \code{TRUE} or \code{FALSE}}

\item{excludeYprobes}{Logical argument.If \code{TRUE}, it will remove the
cell lines. Default to \code{TRUE}. It will return to the expression set. It
removes probeset that belongs to chromosome y}
}
\description{
Automatically load target microarray data and removes Y chromosome features.
The function also loads an annotation matrix.
}
\examples{

#esLoadTarget(type = NULL, load_target_subset = TRUE, excludeYprobes = TRUE)

}
\references{
"An Introduction to Bioconductor's ExpressionSet Class" \cr Seth
Falcon, Martin Morgan, and Robert Gentleman \cr 6 October, 2006; revised 9
February, 2007 \cr
}
\author{
Shahab Asgharzadeh
}
