\name{esLoadU133NBL}
\alias{esLoadU133NBL}
\title{esLoadU133NBL}
\description{
loaded data from U133 NBL
}
\usage{
esLoadU133NBL(type = "tsr", excludeYprobes = TRUE)
}
\arguments{
  \item{type}{this is the type of load, tsr is default}
  \item{excludeYprobes}{logical command for excluding certain probes in Y}
}
%- maybe also 'usage' for other objects documented here.

\author{
Dr. Shahab Asgharzadeh \cr
Janahan Gnanachandran \cr
Anthony Colombo \cr
Alexander Magsarili \cr

}


