\name{esReduce}
\alias{esReduce}
\title{esReduce}
\description{
Create a common gene expressionset
}
\usage{
esReduce(es, annot)
}

\arguments{
  \item{es}{expression set

}
  \item{annot}{is an annotation object (first column is the probe_id)

}
}

\references{
"An Introduction to Bioconductor's ExpressionSet Class" \cr
Seth Falcon, Martin Morgan, and Robert Gentleman \cr
6 October, 2006; revised 9 February, 2007 \cr

}
\author{
Dr. Shahab Asgharzadeh \cr
Janahan Gnanachandran \cr
Anthony Colombo \cr
Alexander Magsarili \cr

}

\examples{
#results <- esReduce(eset, anno_obj)
#results

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
