% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmfunction_orig.R
\name{glmfunction_orig}
\alias{glmfunction_orig}
\title{General Linear Model}
\usage{
glmfunction_orig(y, var1)
}
\arguments{
\item{y}{the vector of data}

\item{var1}{the covariate (n levels) for ANOVA analyses}
}
\description{
Glmfunction calls the in-house "glm" function to include p-value,
chi-squared, means, but does not compare between groups. The calculated
values are only done for a single vector of data.
}
\examples{

#x <- c(2,3,4,7,9,4,10,6,4,2)
#y <- as.factor(c("a","b","c","a","b","c","b","a","c","c"))
#results <- glmfunction_orig(x,y)
#results 
 


}
\author{
Shahab Asgharzadeh
}
\keyword{~kwd1}
\keyword{~kwd2}
