% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_survival.R
\name{qplot_survival}
\alias{qplot_survival}
\title{qplot_survival}
\usage{
qplot_survival(f.frame, f.CI = "default", f.linesize = 1,
  f.shapesize = 2.5, f.shape = 3, kmcolors = kmPalette)
}
\arguments{
\item{f.frame}{data frame}

\item{f.CI}{data frame of confidence interval. Either \code{default},
\code{TRUE}, or \code{FALSE}}

\item{f.linesize}{thickness setting which is set to 1}

\item{f.shapesize}{used for the aesthetic setting argument for the "size" in
ggplot}

\item{f.shape}{used for the aesthetic setting argument for the "shape" in
ggplot}

\item{kmcolors}{color palette}
}
\description{
Draw survival function, needs object survival package, second parameter if
set to \code{TRUE} will display confidence intervals. Use different plotting
commands dependig whether or not strata's are given.
}
\note{
Requires package 'ggplot' and 'survival'.
}
\examples{

#qplot_survival(f.frame, f.CI = "default", f.linesize = 1, f.shapesize = 2.5, f.shape = 3, kmcolors = kmPalette)

}
\author{
Shahab Asgharzadeh
}
\keyword{~kwd2}
