\name{set_tmp_path_data}
\alias{set_tmp_path_data}
\title{Set_Tmp_Path_Data}
\description{This is a function that creates a temporary path directory for retrieving data, behind the scene task function}
\usage{
set_tmp_path_data()
}

\author{
Dr. Shahab Asgharzadeh \cr
Janahan Gnanachandran \cr
Anthony Colombo \cr
Alexander Magsarili \cr

}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
